+++ src/components/worker/weeklyCalendar/AddWorkRequestModal.tsx
@@ -1,26 +1,19 @@
-import React, { useState, useMemo, useRef, useEffect, useCallback } from "react";
+import React, { useEffect } from "react";
 import {
 	View,
-	Modal,
 	StyleSheet,
 	TouchableOpacity,
-	Dimensions,
-	Animated,
 	ActivityIndicator,
 } from "react-native";
 import { Feather } from "@expo/vector-icons";
 import { Text } from "../../common/Text";
 import WheelPicker from "../../common/WheelPicker";
-import type { WheelPickerItem } from "../../common/WheelPicker";
-import { getContracts, getContractDetail } from "../../../api/workerApi";
+import BottomSheetModal from "../../common/BottomSheetModal";
+import useWorkplaces from "../../../hooks/worker/useWorkplaces";
+import useWorkRequestForm from "../../../hooks/worker/useWorkRequestForm";
+import type { PickerTarget } from "../../../hooks/worker/useWorkRequestForm";
 import { colors } from "../../../constants/colors";
 import { formatCurrency } from "../../../utils/format";
-
-interface WorkplaceOption {
-	contractId: number;
-	workplaceName: string;
-	hourlyWage: number;
-}
 
 interface AddWorkRequestModalProps {
 	visible: boolean;
@@ -34,151 +27,27 @@ interface AddWorkRequestModalProps {
 	}) => void;
 }
 
-type PickerTarget =
-	| "workplace"
-	| "date"
-	| "startHour"
-	| "startMinute"
-	| "endHour"
-	| "endMinute"
-	| "breakMinutes"
-	| null;
-
-const { width: SCREEN_WIDTH } = Dimensions.get("window");
-
-// 시간/분 옵션 생성
-const HOUR_ITEMS: WheelPickerItem[] = Array.from({ length: 24 }, (_, i) => ({
-	label: String(i).padStart(2, "0"),
-	value: i,
-}));
-
-const MINUTE_ITEMS: WheelPickerItem[] = Array.from({ length: 60 }, (_, i) => ({
-	label: String(i).padStart(2, "0"),
-	value: i,
-}));
-
-const BREAK_ITEMS: WheelPickerItem[] = Array.from({ length: 7 }, (_, i) => ({
-	label: `${i * 10}`,
-	value: i * 10,
-}));
-
-// 이번 달 날짜 옵션 생성
-const getDateItems = (): WheelPickerItem[] => {
-	const now = new Date();
-	const year = now.getFullYear();
-	const month = now.getMonth();
-	const daysInMonth = new Date(year, month + 1, 0).getDate();
-
-	return Array.from({ length: daysInMonth }, (_, i) => {
-		const day = i + 1;
-		return {
-			label: `${month + 1}/${day}`,
-			value: day,
-		};
-	});
-};
-
 const AddWorkRequestModal: React.FC<AddWorkRequestModalProps> = ({
 	visible,
 	onClose,
 	onSubmit,
 }) => {
-	const [workplaces, setWorkplaces] = useState<WorkplaceOption[]>([]);
-	const [isLoading, setIsLoading] = useState(false);
-	const [selectedWorkplaceId, setSelectedWorkplaceId] = useState<
-		number | null
-	>(null);
-	const [selectedDate, setSelectedDate] = useState<number>(
-		new Date().getDate()
-	);
-	const [startHour, setStartHour] = useState(0);
-	const [startMinute, setStartMinute] = useState(0);
-	const [endHour, setEndHour] = useState(0);
-	const [endMinute, setEndMinute] = useState(0);
-	const [breakMinutes, setBreakMinutes] = useState(0);
-	const [activePicker, setActivePicker] = useState<PickerTarget>(null);
+	const { workplaces, isLoading, fetchWorkplaces, workplacePickerItems } =
+		useWorkplaces();
 
-	// 슬라이드업 애니메이션
-	const slideAnim = useRef(new Animated.Value(0)).current;
-	const overlayAnim = useRef(new Animated.Value(0)).current;
-
-	// 모달 열릴 때 계약 목록 → 상세 조회하여 근무지 목록 구성
-	const fetchWorkplaces = useCallback(async () => {
-		setIsLoading(true);
-		try {
-			const contractsRes = await getContracts();
-			const activeContracts =
-				contractsRes.data?.filter((c) => c.isActive) ?? [];
-
-			const details = await Promise.all(
-				activeContracts.map(async (contract) => {
-					const detailRes = await getContractDetail(contract.id);
-					return {
-						contractId: contract.id,
-						workplaceName:
-							detailRes.data?.workplaceName ?? "알 수 없음",
-						hourlyWage: detailRes.data?.hourlyWage ?? 0,
-					};
-				})
-			);
-
-			setWorkplaces(details);
-		} catch {
-			setWorkplaces([]);
-		} finally {
-			setIsLoading(false);
-		}
-	}, []);
+	const {
+		selectedWorkplaceId,
+		activePicker,
+		handlePickerChange,
+		togglePicker,
+		getPickerConfig,
+		getDisplayValue,
+		buildSubmitData,
+	} = useWorkRequestForm(workplacePickerItems);
 
 	useEffect(() => {
-		if (visible) {
-			fetchWorkplaces();
-			Animated.parallel([
-				Animated.timing(overlayAnim, {
-					toValue: 1,
-					duration: 250,
-					useNativeDriver: true,
-				}),
-				Animated.spring(slideAnim, {
-					toValue: 1,
-					damping: 20,
-					stiffness: 200,
-					useNativeDriver: true,
-				}),
-			]).start();
-		} else {
-			slideAnim.setValue(0);
-			overlayAnim.setValue(0);
-		}
+		if (visible) fetchWorkplaces();
 	}, [visible]);
-
-	const handleClose = () => {
-		Animated.parallel([
-			Animated.timing(overlayAnim, {
-				toValue: 0,
-				duration: 200,
-				useNativeDriver: true,
-			}),
-			Animated.timing(slideAnim, {
-				toValue: 0,
-				duration: 200,
-				useNativeDriver: true,
-			}),
-		]).start(() => {
-			onClose();
-		});
-	};
-
-	const dateItems = useMemo(() => getDateItems(), []);
-
-	const workplaceItems: WheelPickerItem[] = useMemo(
-		() =>
-			workplaces.map((wp) => ({
-				label: wp.workplaceName,
-				value: wp.contractId,
-			})),
-		[workplaces]
-	);
 
 	const selectedWorkplace = workplaces.find(
 		(wp) => wp.contractId === selectedWorkplaceId
@@ -186,128 +55,21 @@ const AddWorkRequestModal: React.FC<AddWorkRequestModalProps> = ({
 	const salary = selectedWorkplace?.hourlyWage ?? 0;
 
 	const handleSubmit = () => {
-		if (!selectedWorkplaceId) return;
-
-		const now = new Date();
-		const year = now.getFullYear();
-		const month = now.getMonth() + 1;
-		const dateStr = `${year}-${String(month).padStart(2, "0")}-${String(selectedDate).padStart(2, "0")}`;
-
-		onSubmit({
-			contractId: selectedWorkplaceId,
-			requestedWorkDate: dateStr,
-			requestedStartTime: `${String(startHour).padStart(2, "0")}:${String(startMinute).padStart(2, "0")}`,
-			requestedEndTime: `${String(endHour).padStart(2, "0")}:${String(endMinute).padStart(2, "0")}`,
-			requestedBreakMinutes: breakMinutes,
-		});
+		const data = buildSubmitData();
+		if (data) onSubmit(data);
 	};
 
-	const handlePickerChange = (value: string | number) => {
-		switch (activePicker) {
-			case "workplace":
-				setSelectedWorkplaceId(value as number);
-				break;
-			case "date":
-				setSelectedDate(value as number);
-				break;
-			case "startHour":
-				setStartHour(value as number);
-				break;
-			case "startMinute":
-				setStartMinute(value as number);
-				break;
-			case "endHour":
-				setEndHour(value as number);
-				break;
-			case "endMinute":
-				setEndMinute(value as number);
-				break;
-			case "breakMinutes":
-				setBreakMinutes(value as number);
-				break;
-		}
-	};
-
-	const getPickerConfig = (): {
-		items: WheelPickerItem[];
-		selectedValue: string | number;
-		width: number;
-	} => {
-		switch (activePicker) {
-			case "workplace":
-				return {
-					items: workplaceItems,
-					selectedValue: selectedWorkplaceId ?? workplaces[0]?.contractId ?? 0,
-					width: SCREEN_WIDTH - 80,
-				};
-			case "date":
-				return {
-					items: dateItems,
-					selectedValue: selectedDate,
-					width: 120,
-				};
-			case "startHour":
-			case "endHour":
-				return {
-					items: HOUR_ITEMS,
-					selectedValue:
-						activePicker === "startHour" ? startHour : endHour,
-					width: 80,
-				};
-			case "startMinute":
-			case "endMinute":
-				return {
-					items: MINUTE_ITEMS,
-					selectedValue:
-						activePicker === "startMinute"
-							? startMinute
-							: endMinute,
-					width: 80,
-				};
-			case "breakMinutes":
-				return {
-					items: BREAK_ITEMS,
-					selectedValue: breakMinutes,
-					width: 80,
-				};
-			default:
-				return { items: [], selectedValue: 0, width: 80 };
-		}
-	};
+	const pickerConfig = getPickerConfig();
 
 	const renderSelectField = (
-		_label: string,
 		target: PickerTarget,
 		style?: object
 	) => {
-		let displayValue = "";
-		switch (target) {
-			case "workplace":
-				displayValue = selectedWorkplace?.workplaceName ?? "근무지를 선택하세요..";
-				break;
-			case "date": {
-				const month = new Date().getMonth() + 1;
-				displayValue = `${month}/${selectedDate}`;
-				break;
-			}
-			case "startHour":
-				displayValue = String(startHour).padStart(2, "0");
-				break;
-			case "startMinute":
-				displayValue = String(startMinute).padStart(2, "0");
-				break;
-			case "endHour":
-				displayValue = String(endHour).padStart(2, "0");
-				break;
-			case "endMinute":
-				displayValue = String(endMinute).padStart(2, "0");
-				break;
-			case "breakMinutes":
-				displayValue = String(breakMinutes);
-				break;
-		}
-
-		const isPlaceholder = target === "workplace" && !selectedWorkplace;
+		const isWorkplace = target === "workplace";
+		const displayValue = isWorkplace
+			? (selectedWorkplace?.workplaceName ?? "근무지를 선택하세요..")
+			: getDisplayValue(target);
+		const isPlaceholder = isWorkplace && !selectedWorkplace;
 
 		return (
 			<TouchableOpacity
@@ -316,9 +78,7 @@ const AddWorkRequestModal: React.FC<AddWorkRequestModalProps> = ({
 					activePicker === target && styles.selectFieldActive,
 					style,
 				]}
-				onPress={() =>
-					setActivePicker(activePicker === target ? null : target)
-				}
+				onPress={() => togglePicker(target)}
 				activeOpacity={0.7}
 			>
 				<Text
@@ -339,183 +99,118 @@ const AddWorkRequestModal: React.FC<AddWorkRequestModalProps> = ({
 		);
 	};
 
-	const pickerConfig = getPickerConfig();
-
-	const screenHeight = Dimensions.get("window").height;
-	const translateY = slideAnim.interpolate({
-		inputRange: [0, 1],
-		outputRange: [screenHeight, 0],
-	});
-
 	return (
-		<Modal
-			visible={visible}
-			transparent
-			animationType="none"
-			onRequestClose={handleClose}
-		>
-			<View style={styles.overlay}>
-				<Animated.View
-					style={[
-						styles.overlayBackground,
-						{ opacity: overlayAnim },
-					]}
-				>
-					<TouchableOpacity
-						style={StyleSheet.absoluteFill}
-						activeOpacity={1}
-						onPress={handleClose}
+		<BottomSheetModal visible={visible} onClose={onClose}>
+			<Text weight="Bold" style={styles.title}>
+				근무 추가 요청하기
+			</Text>
+
+			{/* 근무지 */}
+			<View style={styles.section}>
+				<Text weight="Medium" style={styles.label}>
+					근무지
+				</Text>
+				{isLoading ? (
+					<ActivityIndicator
+						size="small"
+						color={colors.primary}
+						style={{ alignSelf: "flex-start", padding: 10 }}
 					/>
-				</Animated.View>
-				<Animated.View
-					style={[
-						styles.modalContainer,
-						{ transform: [{ translateY }] },
-					]}
-				>
-					<View style={styles.handle} />
-
-					<Text weight="Bold" style={styles.title}>
-						근무 추가 요청하기
-					</Text>
-
-					{/* 근무지 */}
-					<View style={styles.section}>
-						<Text weight="Medium" style={styles.label}>
-							근무지
-						</Text>
-						{isLoading ? (
-							<ActivityIndicator
-								size="small"
-								color={colors.primary}
-								style={{ alignSelf: "flex-start", padding: 10 }}
-							/>
-						) : (
-							renderSelectField("근무지", "workplace", {
-								alignSelf: "flex-start",
-							})
-						)}
-					</View>
-
-					{/* 근무 시간 */}
-					<View style={styles.section}>
-						<Text weight="Medium" style={styles.label}>
-							근무 시간
-						</Text>
-						<View style={styles.timeRow}>
-							{renderSelectField("날짜", "date", {
-								flex: 0,
-								minWidth: 70,
-							})}
-							{renderSelectField("시", "startHour", styles.timeField)}
-							<Text style={styles.timeSeparator}>:</Text>
-							{renderSelectField("분", "startMinute", styles.timeField)}
-							<Text weight="Medium" style={styles.timeTilde}>
-								~
-							</Text>
-							{renderSelectField("시", "endHour", styles.timeField)}
-							<Text style={styles.timeSeparator}>:</Text>
-							{renderSelectField("분", "endMinute", styles.timeField)}
-						</View>
-					</View>
-
-					{/* 휴게 시간 & 시급 */}
-					<View style={styles.rowSection}>
-						<View style={styles.halfSection}>
-							<Text weight="Medium" style={styles.label}>
-								휴게 시간
-							</Text>
-							<View style={styles.unitRow}>
-								{renderSelectField(
-									"분",
-									"breakMinutes",
-									styles.smallField
-								)}
-								<Text weight="Medium" style={styles.unitText}>
-									분
-								</Text>
-							</View>
-						</View>
-						<View style={styles.halfSection}>
-							<Text weight="Medium" style={styles.label}>
-								시급
-							</Text>
-							<View style={styles.unitRow}>
-								<View style={styles.salaryField}>
-									<Text weight="Medium" style={styles.salaryText}>
-										{formatCurrency(salary)}
-									</Text>
-								</View>
-								<Text weight="Medium" style={styles.unitText}>
-									원
-								</Text>
-							</View>
-						</View>
-					</View>
-
-					{/* WheelPicker 영역 */}
-					{activePicker && pickerConfig.items.length > 0 && (
-						<View style={styles.pickerArea}>
-							<View style={styles.pickerWrapper}>
-								<WheelPicker
-									items={pickerConfig.items}
-									selectedValue={pickerConfig.selectedValue}
-									onValueChange={handlePickerChange}
-									width={pickerConfig.width}
-								/>
-							</View>
-						</View>
-					)}
-
-					{/* 요청 보내기 버튼 */}
-					<View style={styles.submitRow}>
-						<TouchableOpacity
-							style={[
-								styles.submitButton,
-								!selectedWorkplaceId && styles.submitButtonDisabled,
-							]}
-							onPress={handleSubmit}
-							activeOpacity={0.8}
-							disabled={!selectedWorkplaceId}
-						>
-							<Feather name="send" size={16} color={colors.white} />
-							<Text weight="Bold" style={styles.submitText}>
-								요청 보내기
-							</Text>
-						</TouchableOpacity>
-					</View>
-				</Animated.View>
+				) : (
+					renderSelectField("workplace", {
+						alignSelf: "flex-start",
+					})
+				)}
 			</View>
-		</Modal>
+
+			{/* 근무 시간 */}
+			<View style={styles.section}>
+				<Text weight="Medium" style={styles.label}>
+					근무 시간
+				</Text>
+				<View style={styles.timeRow}>
+					{renderSelectField("date", {
+						flex: 0,
+						minWidth: 70,
+					})}
+					{renderSelectField("startHour", styles.timeField)}
+					<Text style={styles.timeSeparator}>:</Text>
+					{renderSelectField("startMinute", styles.timeField)}
+					<Text weight="Medium" style={styles.timeTilde}>
+						~
+					</Text>
+					{renderSelectField("endHour", styles.timeField)}
+					<Text style={styles.timeSeparator}>:</Text>
+					{renderSelectField("endMinute", styles.timeField)}
+				</View>
+			</View>
+
+			{/* 휴게 시간 & 시급 */}
+			<View style={styles.rowSection}>
+				<View style={styles.halfSection}>
+					<Text weight="Medium" style={styles.label}>
+						휴게 시간
+					</Text>
+					<View style={styles.unitRow}>
+						{renderSelectField("breakMinutes", styles.smallField)}
+						<Text weight="Medium" style={styles.unitText}>
+							분
+						</Text>
+					</View>
+				</View>
+				<View style={styles.halfSection}>
+					<Text weight="Medium" style={styles.label}>
+						시급
+					</Text>
+					<View style={styles.unitRow}>
+						<View style={styles.salaryField}>
+							<Text weight="Medium" style={styles.salaryText}>
+								{formatCurrency(salary)}
+							</Text>
+						</View>
+						<Text weight="Medium" style={styles.unitText}>
+							원
+						</Text>
+					</View>
+				</View>
+			</View>
+
+			{/* WheelPicker 영역 */}
+			{activePicker && pickerConfig.items.length > 0 && (
+				<View style={styles.pickerArea}>
+					<View style={styles.pickerWrapper}>
+						<WheelPicker
+							items={pickerConfig.items}
+							selectedValue={pickerConfig.selectedValue}
+							onValueChange={handlePickerChange}
+							width={pickerConfig.width}
+						/>
+					</View>
+				</View>
+			)}
+
+			{/* 요청 보내기 버튼 */}
+			<View style={styles.submitRow}>
+				<TouchableOpacity
+					style={[
+						styles.submitButton,
+						!selectedWorkplaceId && styles.submitButtonDisabled,
+					]}
+					onPress={handleSubmit}
+					activeOpacity={0.8}
+					disabled={!selectedWorkplaceId}
+				>
+					<Feather name="send" size={16} color={colors.white} />
+					<Text weight="Bold" style={styles.submitText}>
+						요청 보내기
+					</Text>
+				</TouchableOpacity>
+			</View>
+		</BottomSheetModal>
 	);
 };
 
 const styles = StyleSheet.create({
-	overlay: {
-		flex: 1,
-		justifyContent: "flex-end",
-	},
-	overlayBackground: {
-		...StyleSheet.absoluteFillObject,
-		backgroundColor: "rgba(0, 0, 0, 0.5)",
-	},
-	modalContainer: {
-		backgroundColor: colors.white,
-		borderTopLeftRadius: 24,
-		borderTopRightRadius: 24,
-		paddingTop: 12,
-		paddingHorizontal: 24,
-		paddingBottom: 40,
-		maxHeight: "90%",
-	},
-	handle: {
-		width: 40,
-		height: 4,
-		backgroundColor: colors.grey,
-		borderRadius: 2,
-		alignSelf: "center",
-		marginBottom: 20,
-	},
 	title: {
 		fontSize: 20,
 		color: colors.textPrimary,
